+++ src/components/About/AboutMappedData.tsx
