+++ src/components/About/AboutText.tsx
