+++ src/components/About/index.tsx
@@ -1,7 +1,7 @@
 import uuid from "react-uuid";
 import { AboutMappedData } from "./AboutMappedData";
-import { AboutText } from "./AboutText";
 import React from "react";
+import { AboutText } from "./AboutText";
 
 /* about component */
 function About() {
