+++ src/components/Hero/HeroHeading.tsx
