+++ src/components/Hero/HeroImage.tsx
