+++ src/components/Navbar/NavbarLinksPlatforms.tsx
