+++ src/components/Pricing/PricingCardLeft.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { GetStarted } from "./GetStarted";
-import { mappeddataCheckmark } from "./Pricing";
+import { mappeddataCheckmark } from ".";
+import { GetStarted } from "../GetStarted";
 
 /**
  * create pricing card on the top/left side (based on viewpoint)
