+++ src/components/Pricing/PricingText.tsx
